Refactored Digests and Signatures per feedback I got from before the holiday break
