--- src/main/scala/co/topl/brambl/digests/Hash.scala
@@ -1,8 +1,0 @@
-package co.topl.brambl.digests
-
-import quivr.models._
-
-trait Hash {
-  val routine: String
-  def hash(preimage: Preimage): Digest
-}
