+++ src/main/scala/co/topl/brambl/routines/Routine.scala
@@ -1,0 +1,5 @@
+package co.topl.brambl.routines
+
+trait Routine {
+  val routine: String
+}
