+++ src/main/scala/co/topl/brambl/routines/digests/Hash.scala
@@ -1,0 +1,8 @@
+package co.topl.brambl.routines.digests
+
+import co.topl.brambl.routines.Routine
+import quivr.models._
+
+trait Hash extends Routine{
+  def hash(preimage: Preimage): Digest
+}
