+++ src/main/scala/co/topl/brambl/routines/digests/validators/Blake2b256DigestInterpreter.scala
@@ -1,19 +1,14 @@
-package co.topl.brambl.digests
+package co.topl.brambl.routines.digests.validators
 
+import co.topl.brambl.routines.Routine
 import co.topl.crypto.hash.blake2b256
 import co.topl.quivr.algebras.DigestVerifier
 import co.topl.quivr.runtime.{QuivrRuntimeError, QuivrRuntimeErrors}
 import com.google.protobuf.ByteString
-import quivr.models.Digest
-import quivr.models.DigestVerification
-import quivr.models.Preimage
+import quivr.models.{Digest, DigestVerification, Preimage}
 
-object Blake2b256Digest extends DigestVerifier[Option] with Hash {
+object Blake2b256DigestInterpreter extends DigestVerifier[Option] with Routine {
   override val routine: String = "blake2b256"
-
-  override def hash(preimage: Preimage): Digest = Digest().withDigest32(
-    Digest.Digest32(ByteString.copyFrom(blake2b256.hash(preimage.input.toByteArray ++ preimage.salt.toByteArray).value))
-  )
 
   override def validate(v: DigestVerification): Option[Either[QuivrRuntimeError, DigestVerification]] = {
     val test = blake2b256.hash(v.preimage.get.input.toByteArray ++ v.preimage.get.salt.toByteArray).value
