+++ src/main/scala/co/topl/brambl/routines/signatures/Signing.scala
@@ -1,10 +1,10 @@
-package co.topl.brambl.signatures
+package co.topl.brambl.routines.signatures
 
 import co.topl.brambl.Models.{KeyPair, SigningKey}
+import co.topl.brambl.routines.Routine
 import quivr.models._
 
-trait Signing {
-  val routine: String
+trait Signing extends Routine{
   def createKeyPair(seed: Array[Byte]): KeyPair
   def sign(sk:            SigningKey, msg: SignableBytes): Witness
 }
